WIP: Restore multiple vCPUs.
With this commint, restoring VMs with more than 1 vCPU *seems* to work,
in the sense that we get an interactive shell after restore. However,
most of the vCPUs seem to generate a lot of vmexits with reason 1.

In addition, things like "sleep 5" never come back (last forever).
Needs additional work.