arm64/locore.S: Disable the EL2 MMU
If the MMU is active when booting in EL2 then it is using a page
table where VA == PA. But the page table won't have entries for the
hypervisor EL2 code from arm64/vmm/hyp.S, which will only be
loaded into memory with the vmm module.

So we disable the MMU in EL2 to make the vmm hypervisor code run
successfully.