Fix pci_device not used bug for save/restore
We were wrongly assuming that all pci that had a save/restore
functionality implemented were used. When one of them is not used (e.g.,
virtio-net), the save/restore process would fail. We fixed it by setting
the device size to 0 at snapshot and ignore restoring it if the size is
0.

Add <CR> for some of the messages printed at console.

Move pci devs array to global scope.

Signed-off-by: Elena Mihailescu <elenamihailescu22@gmail.com>