Remove check old and new vioapic id at restore
When trying to migrate a guest that was using pci_fbuf and pci_xhci, the
migration process would fail because of the vioapic restore process: it
was required that the old (saved) vioapic to have the same id as the
new one. When using fbuf + xhci, that values will differ. The values
do not need to be equal because at some point, the guest will set the
vioapic->id field.

Signed-off-by: Elena Mihailescu <elenamihailescu22@gmail.com>