arm64: bhyve: Make CNTV_{CVAL, CTL}_EL0 part of the vcpu context
Two different VMs can run on the same CPU. Save and restore the virtual
timer registers registers when doing a world switch so the VMs don't
interfere with each other.

Note that this is done in an inefficient manner. The registers need to
be saved/restored only when switch between **different** VMs, not when
switching between the host and the **same** VM.