arm/arm64: Enable Virtual Host Extensions if available
Enable Virtual Host Extensions (VHE) for running in EL2 instead of EL1.
When FreeBSD will implement virtualization on arm64, VHE will make the
hypervisor simpler and faster.

Virtual Host Extensions, when enabled, does the following:
(1) Redirects all accesses from the EL1 register to the corresponding EL2
    register.
(2) New registers become accessible at EL2 to match the EL1 registers, and
    some existing registers are modified to have the same layout as the
    EL1 registers.

This allows FreeBSD on arm64 to run in EL2 with minimal changes, which are
localized in the machine-dependent boot code.

The only modification to the arm architecture is made to the generic
timer driver, which is shared with arm64. On arm64 with VHE enabled,
the driver will set up the EL2 timers interrupt handlers. This is
necessary because all accesses to the timer *_EL0 registers are
redirected to the corresponding *_EL2 registers.