Add functionality for live migration
1. vm_get_dirty_page_list searches for the dirty pages in the guest assigned
vm_objects (one for lowmem and one for highmem). The list contains the
dirty pages indexes.

2. vm_copy_vmm_pages copies guest's related vm_pages from vm_objects
to buffer (in live migration, it is used to get the dirty pages'
content and to send it to the destination host) or from buffer to
vm_objects (in live migration, it is used to set the destination guest
memory to be identical with the source guest memory).

3. add IOCTLs to call the vm_get_dirty_page_list and vm_copy_vmm_pages
functions

Signed-off-by: Elena Mihailescu <elenamihailescu22@gmail.com>